deletion_detector/requirements.txt
pyrogram==2.0.106
aiohttp==3.9.1
python-dotenv==1.0.0
TgCrypto==1.2.5